backend/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
redis==5.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8

# Core dependencies (compatible with Python 3.12)
pydantic==2.5.0
pydantic-settings==2.0.3
httpx==0.25.2
requests==2.31.0

# Machine Learning & AI (optional for basic functionality)
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
openai>=1.0.0

# Financial APIs
plaid-python==9.1.0

# Optional ML dependencies
torch>=2.0.0; python_version>="3.8"
transformers>=4.30.0; python_version>="3.8"

# Time series forecasting (optional)
prophet>=1.1.0; python_version>="3.8"

# Monitoring & Logging (optional)
sentry-sdk[fastapi]>=1.30.0; python_version>="3.8"
structlog>=23.0.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
